backend/monitor_notification_changes/requirements.txt
firebase-admin==6.2.0
functions-framework==3.4.0
google-cloud-tasks==2.13.1
google-cloud-firestore==2.11.1
requests==2.31.0
google-cloud-secret-manager==2.16.4
google-events-cloud==0.5.2